feat: Use VehicleSummaryViewModel in Index view
Change the Index to use VehicleSummaryViewModel instead of ParkedVehicle as its model.

This requires adding ParkedVehicleId as a property of the ViewModel so that the Edit/Details/Delete links can link to the proper pages.

The ParkedTimeAmount displays oddly, going out to fractions of a second. The formatting of this and of the displayed names should be changed in design.