refactor: Add ParkingEnded intermediate view
Add a ParkingEnded view that is returned as soon as EndParking is verified. This view takes a ReceiptViewModel instance generated by EndParking's Post method.

ParkingEnded also has a Post method that re-constructs the original ReceiptViewModel and then passes it to the Receipt view. This is necessary as at this point, the vehicle has already been removed from the database.